reworked code to get student page up and running

+++ api/providers/StudentProvider.js
@@ -15,19 +15,33 @@ export class StudentProvider{
     }
 
     getStudent(id){
+        if (typeof id === 'undefined')
+            throw new Error("ArgumentError: id cannot be undefined!");
         return this._model.findOne({_id: id});
     }
 
     postStudent(studentData){
+        if (typeof studentData === 'undefined')
+            throw new Error("ArgumentError: student data is undefined!");
         return new this._model(studentData).save();
     }
 
     updateStudent(id, studentData){
+        if (typeof id === 'undefined' || typeof studentData === 'undefined')
+            throw new Error("ArgumentError: id and/or student data is undefined!");
         return this._model.findOneAndUpdate({_id: id}, studentData, {new: true});
     }
 
     deleteStudent(id){
+        if (typeof id === 'undefined')
+            throw new Error("ArgumentError: id cannot be undefined!");
         return this._model.findOneAndRemove({_id: id});
+    }
+
+    deleteStudentsById(ids){
+        if (typeof ids === 'undefined')
+            throw new Error("ArgumentError: ids cannot be undefined!");
+        return this._model.deleteMany({ _id: { $in: ids } });
     }
 }
